+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Added method to sign a message
 - Added method to sign Typed Data (EIP-712)
 - Added importWallet() to import account using privateKey.
+- Added sign() to sign a message or transaction and get signature along with v,r,s.